+++ source/steps/build/generate_revision_file.sh
@@ -1,29 +1,11 @@
 #!/usr/bin/env bash
 
-function generate_revision_file_ensure_var_exists
-{
-    for defined_variable_name in "VERBOSE" "VERY_VERBOSE" "DEBUG"
-    do
-        if ! [[ -v "$defined_variable_name" ]]
-        then
-            VERBOSE=
-            VERY_VERBOSE=
-            DEBUG=
-            error "Something unexpected happened: $defined_variable_name should be defined"
-
-            return 1
-        fi
-    done
-
-    return 0
-}
-readonly -f "generate_revision_file_ensure_var_exists"
-
 function generate_revision_file
 {
     do_not_run_twice || return $?
-    ${VERBOSE} && display_title 1 "Generating .REVISION file"
-    generate_revision_file_ensure_var_exists || return $?
+    reset_title_level
+    display_title "Generating .REVISION file"
+    increase_title_level
 
     declare -r workspace="$1"
     declare -r revision="$2"
@@ -32,20 +14,17 @@ function generate_revision_file
     declare -r output_file="$(mktemp -t deploy.XXXXXXXXXX)"
     declare -i return_code=0
 
-    cd "$repository_path"
-    git --no-pager show --quiet "$revision" > "$file" 2>"$output_file" || {
-        return_code=1
-        error "Unable to create revision file $file."
+    is_debug && display_title "\e[33mgit --git-dir=\"$repository_path\" --no-pager show --quiet \"$revision\"\e[39m"
 
-        >&2 printf 'Following is the output of the command\n'
-        >&2 printf '######################################\n'
-        cat "$output_file"
+    git --git-dir="$repository_path" --no-pager show --quiet "$revision" >"$file" 2>"$output_file" || {
+        return_code=1
+        error_with_output_file "$output_file" "Unable to create revision file $file."
     }
 
-    rm "$output_file"
-    cd "$OLDPWD"
+    [[ -f "$output_file" ]] && rm "$output_file"
+
+    decrease_title_level
 
     return ${return_code}
 }
-
 readonly -f "generate_revision_file"
